net: use non-variadic tuple-based future in accept()
Since variadic futures are now deprecated, the accept() function
is transformed to use tuple-based futures and thus avoid warnings.

Since it's a minor API change, it's possible to maintain compatibility
by keeping server_socket::accept() variadic, but providing a newer
server_socket::accept2() and using it in httpd and other code.
Moreover, the deprecated variadic accept() will be implemented
as a thin layer on top of accept2().

If that's the preferred way, a second version of this patch can be
prepared.