semaphore: add named semaphore
By default, semaphore exceptions bring along very little context:
either that *a* semaphore was broken or that it timed out.
In order to make debugging easier without introducing significant
runtime costs, a notion of named semaphore is added.
A named semaphore is simply a semaphore with statically defined
name, which is present in its errors, bringing valuable context.
A semaphore defined as:
  auto sem = semaphore(0);
will present the following message when it breaks:
  "Semaphore broken"
However, a named semaphore:
  struct name {
    static constexpr const char* value = "io_concurrency_sem";
  };
  auto named_sem = named_semaphore<name>(0);
will present a message with at least some debugging context:
  "Semaphore broken: io_concurrency_sem"
It's not much, but it would really help in pinpointing bugs
without having to inspect core dumps.

At the same time, it does not incur any costs for normal
semaphore operations, but instead only uses more CPU in case
an error is actually thrown, which is considered rare
and not to be on the hot path.

Tests: unit(dev)