Kafka TCP connection class
This change adds a basic, low level network connection management
for the Kafka client. Later, this should be encapsulated in a TCP Client class
to work on the Kafka messages abstraction.
So far the tests are also somewhat makeshift - ideally, they have to be separated
into both unit and integration tests, with the former not rellying on existing
Kafka brokers.