Add Connection Manager class
This change adds a connection manager unit to keep track of open connections
and allow for a fast access by the producer. With that we can now manage
multiple brokers and read their addresses from metadata requests.
It also introduces a dummy partitioner for further development.

Connection Pool + Multibroker demo